factor out common version
