+++ src/daft-core/src/prelude.rs
@@ -18,10 +18,6 @@ pub use daft_schema::schema::{Schema, SchemaRef};
 // Re-export join-related types
 pub use crate::join::{JoinStrategy, JoinType};
 
-// Re-export version information
-pub use crate::DAFT_BUILD_TYPE;
-pub use crate::VERSION;
-
 // You might want to include a glob import for users who want everything
 pub mod all {
     pub use super::*;
