+++ src/daft-schema/Cargo.toml
@@ -4,6 +4,7 @@ common-arrow-ffi = {path = "../common/arrow-ffi", default-features = false}
 common-display = {path = "../common/display", default-features = false}
 common-error = {path = "../common/error", default-features = false}
 common-py-serde = {path = "../common/py-serde", default-features = false}
+common-version = {path = "../common/version", default-features = false}
 html-escape = {workspace = true}
 indexmap = {workspace = true}
 num-derive = {workspace = true}
